Fixed access to AddTimer() error. Created mat_pink material for PowerUps. Added Oscillation to PowerUps
Fixed the reference to the Timer class to call AddTime(). Created material mat_pink and added it to the PowerBall prefab. Also created an Oscillation class to give movement on the y axis (script from Black Bug on YouTube edited for my purpose) and added the script to the PowerBall prefab. Adjusted the amount of time added with PowerUp.